[#21] Rename/Style: 로직의 역할에 따른 파일구조/파일명 변경 및 함수명 변경
router와 loader의 역할이 라우팅이라는 하나의 관심사라고 생각하여 routes 폴더를 생성하여 하위에서 관리하도록 변경하였습니다.
signinLoader 함수가 signup에도 사용되므로 범용성을 가지는 명칭인 authLoader로 변경하였습니다.
