+++ src/App.js
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router';
-import { router } from './router';
+import { router } from './routes/router';
 import './App.css';
 
 const App = () => {
