+++ src/pages/index.js
@@ -2,5 +2,3 @@ export * from './SignIn';
 export * from './SignUp';
 export * from './Todo';
 export * from './Error';
-export { signInLoader } from './SignIn/loader';
-export { todoLoader } from './Todo/loader';
