+++ src/routes/loaders/authLoader.js
@@ -2,7 +2,7 @@ import { redirect } from 'react-router-dom';
 
 import { USER_TOKEN_KEY } from 'constants';
 
-export const signInLoader = () => {
+export const authLoader = () => {
   const userToken = localStorage.getItem(USER_TOKEN_KEY);
   if (userToken) {
     return redirect('/todo');
